move stream into it's own file
